:recycle: Rename intro screen and refactor to use navigation and translation hooks
