+++ src/components/screens/IntroScreen/IntroScreen.js
@@ -2,16 +2,21 @@
 import React from 'react';
 import { View } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { useNavigation } from '@react-navigation/native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { translate } from 'react-i18next';
+import i18next from 'i18next';
+
 import activityHistoryImg from 'assets/images/intro/activityHistory3x.png';
 import tokensTransferImg from 'assets/images/intro/tokensTransfer3x.png';
 import secureAuthenticationImg from 'assets/images/intro/secureAuthentication3x.png';
 import easyAccessImg from 'assets/images/intro/easyAccess3x.png';
 import Slider from '../../shared/Slider';
-import styles from './styles';
 
-const Intro = ({ navigation, t }) => {
+import styles from './IntroScreen.styles';
+
+export default function IntroScreen() {
+  const navigation = useNavigation();
+
   const skip = () => {
     AsyncStorage.setItem('@lisk-mobile-intro', 'true');
     navigation.push('AuthMethod', { signOut: true });
@@ -20,28 +25,32 @@ const Intro = ({ navigation, t }) => {
   const slides = [
     {
       step: 1,
-      title: t('Activity history'),
-      description: t(
+      title: i18next.t('Activity history'),
+      description: i18next.t(
         'Get a full overview of your current balance, transaction history and much more.'
       ),
       imageSrc: activityHistoryImg,
     },
     {
       step: 2,
-      title: t('Token transfer'),
-      description: t('Transfer your tokens easily to other accounts, by simply scanning QR Codes.'),
+      title: i18next.t('Token transfer'),
+      description: i18next.t(
+        'Transfer your tokens easily to other accounts, by simply scanning QR Codes.'
+      ),
       imageSrc: tokensTransferImg,
     },
     {
       step: 3,
-      title: t('Secure authentication'),
-      description: t('Access all functions via advanced biometric authentication.'),
+      title: i18next.t('Secure authentication'),
+      description: i18next.t('Access all functions via advanced biometric authentication.'),
       imageSrc: secureAuthenticationImg,
     },
     {
       step: 4,
-      title: t('Easy access'),
-      description: t('Create an account using passphrase to access your LSK cryptocurrency.'),
+      title: i18next.t('Easy access'),
+      description: i18next.t(
+        'Create an account using passphrase to access your LSK cryptocurrency.'
+      ),
       imageSrc: easyAccessImg,
       acceptTermsSwitch: true,
     },
@@ -54,5 +63,4 @@ const Intro = ({ navigation, t }) => {
       </View>
     </SafeAreaView>
   );
-};
-export default translate()(Intro);
+}
