+++ src/components/screens/IntroScreen/IntroScreen.styles.js
